refactor(input/gui): move guiform to standalone package
